src/test/java/kasutajaliidesedApi/utils/FileUtilsTest.java
package kasutajaliidesedApi.utils;

import org.easymock.EasyMockRunner;
import org.junit.Test;
import org.junit.runner.RunWith;

import java.io.File;
import java.io.IOException;
import java.io.InputStream;
import java.io.PrintWriter;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertNull;

@RunWith(EasyMockRunner.class)
public class FileUtilsTest {

    @Test
    public void getFileAsStreamNoFile() {
        InputStream inputStream = FileUtils.getFileAsStream("wrong.xml");
        assertNull(inputStream);
    }

    @Test
    public void getFileAsStream() throws IOException {
        // create tmp file
        File tempFile = File.createTempFile("testFileForGetFileAsStrem", ".kasutajaliidesedApi");
        PrintWriter writer = new PrintWriter(tempFile, "UTF-8");
        writer.println("bla bla bla");
        writer.close();
        tempFile.deleteOnExit();

        InputStream inputStream = FileUtils.getFileAsStream(tempFile.getAbsolutePath());
        assertNotNull(inputStream);
    }
}
